+++ sync-for-expenses/src/codatsyncexpenses/models/operations/initiate_sync.py
@@ -25,7 +25,7 @@ class InitiateSyncResponse:
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""If model is incorrect"""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     sync_initiated: Optional[shared_syncinitiated.SyncInitiated] = dataclasses.field(default=None)