+++ sync-for-expenses/src/codatsyncexpenses/models/shared/customer.py
@@ -2,14 +2,35 @@
 
 from __future__ import annotations
 import dataclasses
+from ..shared import accountingaddresstype as shared_accountingaddresstype
 from ..shared import contact as shared_contact
 from ..shared import customerstatus as shared_customerstatus
-from ..shared import items as shared_items
 from ..shared import metadata as shared_metadata
 from ..shared import supplementaldata as shared_supplementaldata
 from codatsyncexpenses import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
+
+
+@dataclass_json(undefined=Undefined.EXCLUDE)
+@dataclasses.dataclass
+class CustomerAccountingAddress:
+    type: shared_accountingaddresstype.AccountingAddressType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    r"""The type of the address"""
+    city: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('city') }})
+    r"""City of the customer address."""
+    country: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('country') }})
+    r"""Country of the customer address."""
+    line1: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('line1') }})
+    r"""Line 1 of the customer address."""
+    line2: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('line2') }})
+    r"""Line 2 of the customer address."""
+    postal_code: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('postalCode') }})
+    r"""Postal code or zip code."""
+    region: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('region') }})
+    r"""Region of the customer address."""
+    
+
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -25,7 +46,7 @@ class Customer:
     """
     status: shared_customerstatus.CustomerStatus = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('status') }})
     r"""Status of customer."""
-    addresses: Optional[List[shared_items.Items]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('addresses') }})
+    addresses: Optional[List[CustomerAccountingAddress]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('addresses') }})
     r"""An array of Addresses."""
     contact_name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactName') }})
     r"""Name of the main contact for the identified customer."""