+++ sync-for-expenses/src/codatsyncexpenses/models/shared/supplier.py
@@ -2,34 +2,13 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import accountingaddresstype as shared_accountingaddresstype
+from ..shared import items as shared_items
 from ..shared import metadata as shared_metadata
 from ..shared import supplementaldata as shared_supplementaldata
 from ..shared import supplierstatus as shared_supplierstatus
 from codatsyncexpenses import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
-
-
-@dataclass_json(undefined=Undefined.EXCLUDE)
-@dataclasses.dataclass
-class SupplierAccountingAddress:
-    type: shared_accountingaddresstype.AccountingAddressType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
-    r"""The type of the address"""
-    city: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('city') }})
-    r"""City of the customer address."""
-    country: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('country') }})
-    r"""Country of the customer address."""
-    line1: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('line1') }})
-    r"""Line 1 of the customer address."""
-    line2: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('line2') }})
-    r"""Line 2 of the customer address."""
-    postal_code: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('postalCode') }})
-    r"""Postal code or zip code."""
-    region: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('region') }})
-    r"""Region of the customer address."""
-    
-
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -43,7 +22,7 @@ class Supplier:
     """
     status: shared_supplierstatus.SupplierStatus = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('status') }})
     r"""Status of the supplier."""
-    addresses: Optional[List[SupplierAccountingAddress]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('addresses') }})
+    addresses: Optional[List[shared_items.Items]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('addresses') }})
     r"""An array of Addresses."""
     contact_name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactName') }})
     r"""Name of the main contact for the supplier."""