+++ sync-for-expenses/RELEASES.md
@@ -677,3 +677,13 @@ Based on:
 - [python v4.0.0] sync-for-expenses
 ### Releases
 - [PyPI v4.0.0] https://pypi.org/project/codat-sync-for-expenses/4.0.0 - sync-for-expenses
+
+## 2023-10-30 01:25:48
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0] sync-for-expenses
+### Releases
+- [PyPI v4.1.0] https://pypi.org/project/codat-sync-for-expenses/4.1.0 - sync-for-expenses